+++ sdk/python/pulumi_azure/lb/nat_rule.py
@@ -10,7 +10,7 @@ from .. import utilities, tables
 
 class NatRule(pulumi.CustomResource):
     backend_ip_configuration_id: pulumi.Output[str]
-    backend_port: pulumi.Output[int]
+    backend_port: pulumi.Output[float]
     """
     The port used for internal connections on the endpoint. Possible values range between 1 and 65535, inclusive.
     """
@@ -23,7 +23,7 @@ class NatRule(pulumi.CustomResource):
     """
     The name of the frontend IP configuration exposing this rule.
     """
-    frontend_port: pulumi.Output[int]
+    frontend_port: pulumi.Output[float]
     """
     The port for the external endpoint. Port numbers for each Rule must be unique within the Load Balancer. Possible values range between 1 and 65534, inclusive.
     """
@@ -52,10 +52,10 @@ class NatRule(pulumi.CustomResource):
         
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
-        :param pulumi.Input[int] backend_port: The port used for internal connections on the endpoint. Possible values range between 1 and 65535, inclusive.
+        :param pulumi.Input[float] backend_port: The port used for internal connections on the endpoint. Possible values range between 1 and 65535, inclusive.
         :param pulumi.Input[bool] enable_floating_ip: Enables the Floating IP Capacity, required to configure a SQL AlwaysOn Availability Group.
         :param pulumi.Input[str] frontend_ip_configuration_name: The name of the frontend IP configuration exposing this rule.
-        :param pulumi.Input[int] frontend_port: The port for the external endpoint. Port numbers for each Rule must be unique within the Load Balancer. Possible values range between 1 and 65534, inclusive.
+        :param pulumi.Input[float] frontend_port: The port for the external endpoint. Port numbers for each Rule must be unique within the Load Balancer. Possible values range between 1 and 65534, inclusive.
         :param pulumi.Input[str] loadbalancer_id: The ID of the Load Balancer in which to create the NAT Rule.
         :param pulumi.Input[str] name: Specifies the name of the NAT Rule.
         :param pulumi.Input[str] protocol: The transport protocol for the external endpoint. Possible values are `Udp`, `Tcp` or `All`.
