+++ sdk/python/pulumi_azure/operationalinsights/analytics_workspace.py
@@ -29,7 +29,7 @@ class AnalyticsWorkspace(pulumi.CustomResource):
     """
     The name of the resource group in which the Log Analytics workspace is created. Changing this forces a new resource to be created.
     """
-    retention_in_days: pulumi.Output[int]
+    retention_in_days: pulumi.Output[float]
     """
     The workspace data retention in days. Possible values range between 30 and 730.
     """
@@ -58,7 +58,7 @@ class AnalyticsWorkspace(pulumi.CustomResource):
         :param pulumi.Input[str] location: Specifies the supported Azure location where the resource exists. Changing this forces a new resource to be created.
         :param pulumi.Input[str] name: Specifies the name of the Log Analytics Workspace. Workspace name should include 4-63 letters, digits or '-'. The '-' shouldn't be the first or the last symbol. Changing this forces a new resource to be created.
         :param pulumi.Input[str] resource_group_name: The name of the resource group in which the Log Analytics workspace is created. Changing this forces a new resource to be created.
-        :param pulumi.Input[int] retention_in_days: The workspace data retention in days. Possible values range between 30 and 730.
+        :param pulumi.Input[float] retention_in_days: The workspace data retention in days. Possible values range between 30 and 730.
         :param pulumi.Input[str] sku: Specifies the Sku of the Log Analytics Workspace. Possible values are `Free`, `PerNode`, `Premium`, `Standard`, `Standalone`, `Unlimited`, and `PerGB2018` (new Sku as of `2018-04-03`).
         :param pulumi.Input[dict] tags: A mapping of tags to assign to the resource.
         """
